Revise background functions API.
We no longer have an explicit dependency on Gson. Instead, background
functions either receive the JSON payload as a String, or they specify
the Java class into which it should be deserialized. We do this last
part using Gson.

Also avoid deprecated Gson methods, and fix some incorrect code in the
code samples in the javadoc.