Switch from Commons CLI to JCommander.
JCommander has a simpler API and better documentation. It is slightly more code
(this change adds about 37K to the invoker jar).

Also reorganize the Invoker class with a more usual structure.

As a consequence of this change, command line options are spelled e.g. --port
rather than -port. JCommander would support either (or indeed both) but this
is more in line with what people expect.

PiperOrigin-RevId: 293053482