Update Maven dependencies for Java FF.
This is the latest version of each dependency, as reported by `mvn versions:display-dependency-updates`, except for `javax.servlet:javax.servlet-api` where we don't want to go past 3.1 and `org.eclipse.jetty:jetty-*`, where we can go to a later 9.4 but not to 10. In both cases it is because the Servlet API has moved packages in ways that we don't want to have to deal with right now.

Also change some assertEquals calls from JUnit 5, where we are using JUnit 4 everywhere else. These assertEquals calls had their parameters in the wrong order, which is a problem we can avoid by switching to Truth.

PiperOrigin-RevId: 293211216