Restructure the way new-style functions are loaded.
This shares more code between the HTTP and Background flavours,
and improves the logic for guessing the flavour when it is not specified.
It also gives a better exception when the named target class does not exist.

Also replace some uses of deprecated methods in one of the test functions.

PiperOrigin-RevId: 293924171