Initial version of a Maven plugin to run functions.
When appropriately configured in pom.xml, you can use `mvn function:run` to
compile your function and start a web server that serves it.

The version of the plugin artifact will track the version of the Invoker (aka Functions Framework). Since the plugin depends on and runs the Invoker, any time we change the latter we need to change the former.

Further documentation to follow.

PiperOrigin-RevId: 294293170