Make Functions Framework use StackDriver-compatible logging when on GCF.
In this environment, outputting a particular JSON-formatted string will produce a log entry with fields parsed from that string.

Remove the special logging.properties file. The intent was to omit information like the timestamp that would be duplicated by the StackDriver logging framework. But we are now replacing the logger completely, so this logging.properties is no longer consulted.

The logging.properties that was deleted also didn't include the
[`%6$s` field](https://docs.oracle.com/en/java/javase/11/docs/api/java.logging/java/util/logging/SimpleFormatter.html#format(java.util.logging.LogRecord))
which meant that if there was an exception associated with the log message it did not appear.

PiperOrigin-RevId: 302709439