Keep ClassNotFoundException as cause when reporting function not found. (#26)
The code is complicated by the support for legacy "ClassName.method" way of
specifying function targets, which we will delete when we delete java8 support
from GCF. Under the assumption that the user did mean to use the new form, we
remember what the ClassNotFoundException was that caused us to fail, and we
include that in the cause chain of the final exception. That way the user can
see a possible ExceptionInInitializerError or the like that might be the
underlying reason for the failure.