Rename NewFoo to Foo.
Previously we had `BackgroundFunction` and `NewBackgroundFunction`, where the former was the original design where functions were methods that needed to have a particular signature and the latter was the newer design where functions are classes that implement a given interface. We retired `BackgroundFunction` etc a while ago, so we can rename `NewBackgroundFunction` to just `BackgroundFunction`.