Add logic to convert from Legacy Events to CloudEvents.
This allows received Legacy Events to be handled by function code that expects a
CloudEvent input. The translation logic and tests are based closely on the
corresponding code in https://github.com/GoogleCloudPlatform/functions-framework-dotnet.

The new code also passes the relevant conformance tests in
https://github.com/GoogleCloudPlatform/functions-framework-conformance, once
https://github.com/GoogleCloudPlatform/functions-framework-conformance/pull/30
is applied. For now I'm running those tests manually. I plan at least to commit
the script and test functions that allow me to do so.