Add more tests on the data of translated CloudEvents.
These tests check what is in the `data` field of the CloudEvents that we
translate from legacy events. For now they are not all that useful, because that
field is basically just a copy of the JSON payload of the legacy event. But we
do rearrange things slightly for PubSub events, and we are likely to do so for
Firebase events too in the future. Furthermore, if we eventually have a library
of classes representing the various event payloads, then we can use that library
instead of working with the JSON directly. That is what
google-cloudevents-dotnet does, for example.