unblock local development with the pubsub emulator
Currently, the functions frameworks are dependent on some private
dataplane event marshalling logic in order to correctly pass PubSub
events to background functions. This commit implements the same
marshalling logic in the FF in order to enable local development using
the PubSub emulator.

We have already made this change in other languages:
https://github.com/GoogleCloudPlatform/functions-framework-nodejs/pull/272
https://github.com/GoogleCloudPlatform/functions-framework-ruby/pull/100
https://github.com/GoogleCloudPlatform/functions-framework-python/pull/121
https://github.com/GoogleCloudPlatform/functions-framework-go/pull/70