Add conformance tests for background events (#97)
Previously we had been running conformance tests for background
function signature types. This enables these tests in the conformance
github action.

Running the tests uncovered one bug in which JSON fiels with null values
were being dropped from the data param passed to RawBackgroundFunctions.
Fixing this is a non-breaking change so I bumped the minor verison of
the maven package.