fix: make user function exceptions log level SEVERE
The FF is inconsistent in how it logs errors in the user function
between HTTP functions and Event-driven functions. This change makes it
so that all functions log the same error with the same severity if
the user function throws an exception.

Log level SEVERE is required to be consistent with the Cloud Functions &
Error Reporting
[docs](https://cloud.google.com/error-reporting/docs/setup/nodejs#cloud_functions).