chore(implementation)!: use Jetty-12 core without servlets
Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.   Several optimizations are included, including the optimization that small writes can be buffered in such a way to avoid chunked responses when possible.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.