fix: correct Cloud Event retry functionality (#326)
* fix: Correct exception handling for CloudEvents

A previous change resulted broke exception handling for cloud events so
that an exception from the function would not result in a failure
response code. Correct error handling for CloudEvents and add a test to
cover this case to avoid future breakages.

* fix: Do not add execution_id field if logging is disabled

Fix logger to not add field if logging is disabled.