fix: several minor fixes (#341)
* fix: several minor fixes

Changed a couple of maven plugins's scope to provided as recommended by
Maven.

Fixed #117: during first deploy unless the allowUnauthenticated property
was set to true, the gcloud command was waiting on a prompt that was not
visible and could not be interacted with. This change removes the
allowUnauthenticated property's default value and relies on gcloud's
behavior which will default this value to false if prompting is turned
off with '--quiet'. If the allowUnauthenticated is explicitly set, it
will correctly append the --[no-]allow-unauthenticated flag to the
command.

If the gcloud command fails, the plugin will not report a successful
build.

* fix formatting