chore(implementation): use Jetty-12.1 core without servlets (#333)
* chore(implementation)!: use Jetty-12 core without servlets

Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>

* chore(implementation)!: use Jetty-12 core without servlets

Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>

* chore(implementation)!: use Jetty-12 core without servlets

Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>

* chore(implementation)!: use Jetty-12 core without servlets

Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>

* chore(implementation)!: use Jetty-12 core without servlets

Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>

* chore(implementation)!: use Jetty-12 core without servlets

Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>

* chore(implementation)!: use Jetty-12 core without servlets

Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>

* chore(implementation)!: use Jetty-12 core without servlets

Port the invoker to upgrade to Eclipse Jetty-12 version 12.  Specifically using the new core APIs of Eclipse Jetty-12 that allow the overhead of a Servlet container to be avoided.

BREAKING CHANGE: use Java 17 or above, as required by Eclipse Jetty-12.

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>

---------

Signed-off-by: Lachlan Roberts <lachlan.p.roberts@gmail.com>